tests: refactor integration test package
This patch splits up the integration test package down into several
parts, the core integration helpers, the buildkit test workers and
various test helpers.

With this, we can split the worker features off as well, which improves
the general reusability of the package, so we can create our own custom
features in buildx where we reuse this package.

Signed-off-by: Justin Chadwell <me@jedevc.com>