lint: fix linting rules for logrus
The regexp should only match *symbols*, so we need to remove the opening
bracket of the function call. Additionally, we need to use anchors to
match the whole string, so as to not break on logrus.TraceLevel and
similar.

Finally, the error message is reworded to also mention bklog.L.

Signed-off-by: Justin Chadwell <me@jedevc.com>