progressui: modify NewDisplay to accept io.Writer instead of console.File
The original intention of accepting `console.File` was that the current
uses of this function in the codebase were always files and it seemed
helpful to restrict the API to `console.File` to aid developers into
passing the correct output.

There are circumstances outside of the codebase where the `io.Writer`
versions are used so `NewDisplay` has now been adapted to take an
`io.Writer` to allow for those cases more easily.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>