Skip export of caches with no layers to OCI structures
These happen when a container image is built using only metadata
actions, and hence does not generate any filesystem changes.

Such a degenerate cache is not interesting, and some OCI registries
reject OCI images with no layers.

Specifically, this only affects the `registry` and `local` exporters.

The `inline` exporter already early-outs in this case with the same
warning, and the `gha`, `s3`, and `azblob` exporters are unchanged.

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>