buildkitd: use default config file location when run as root in rootless
When running buildkitd in a rootless environment but as root, buildkitd
misinterprets how it is being run and loads the config file from the
home directory instead of the global location in `/etc/buildkitd`.

This happens when buildkit is being run in docker and docker itself is
being run as rootless. The buildkit daemon sees the user remapping and
assumes that the remapping belongs to itself rather than to docker.

This causes it to load the wrong configuration file as buildkitd is
still being run as "root" but a remapped root that docker created.

Signed-off-by: Jonathan A. Sternberg <jonathan.sternberg@docker.com>