containerd: support custom shim path
Like CRI, we should support custom shim paths. We can just add this as a
new config field "path" under the runtime object.

While "name" can be either the name of a custom shim (from which a path
is derived), or a path itself, if configured as a path, no options can
be provided - this is because to derive the type for the options struct,
we need the name to be a member of a well-known set (currently runc and
hcs).

With this patch, it's now possible to configure a runtime with a custom
shim at a non-default path, and include it's options (which was
previously not possible to do in buildkit).

Signed-off-by: Justin Chadwell <me@jedevc.com>